service-analytics/requirements.txt
certifi
charset-normalizer
contourpy
cycler
filelock
fonttools
fsspec
huggingface-hub
idna
Jinja2
kiwisolver
loguru
MarkupSafe
matplotlib
mpmath
networkx
numpy
packaging
pandas
pillow
psycopg2-binary
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
safetensors
six
sympy
tokenizers
torch
tqdm
transformers
typing_extensions
tzdata
urllib3

# Web Framework Dependencies
Flask
Flask-SocketIO
APScheduler

# Modern AI Dependencies
accelerate
bitsandbytes
peft
trl
sentence-transformers
rank-bm25
scikit-learn
scipy
datasets
evaluate
optimum

# NATS Dependencies
nats-py
asyncio-nats-client

# Vendor API Dependencies
boto3
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
plaid-python